backend/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
anthropic>=0.49.0
pinecone-client==2.2.1
python-dotenv==1.0.0
requests==2.31.0
openai>=1.0.0
nltk
Pillow
python-multipart